[171. Excel Sheet Column Number][Accepted]committed by LeetCode Extension

+++ 171-Excel-Sheet-Column-Number/solution.java
@@ -2,7 +2,7 @@ public class Solution {
     public int titleToNumber(String s) {
         int result = 0;
         for(int i = 0; i < s.length(); i++) {
-            result = result * 26 + s.charAt(i) - 'A' + 1;
+            result = result * 26 + (int)(s.charAt(i) - 'A') + 1;
         }
         return result;
     }
